DatePicker: refactor validation system
